Extract expression pushdown tests to BaseJdbcConnectorTest
This also renames testIsNullPredicatePushdown to
testIsNullExpressionPredicatePushdown to clarify that expression
pushdown of IS NULL is being tested.
This also removes a mistaken assertion that IS NULL cannot be pushed
down if SUPPORTS_PREDICATE_EXPRESSION_PUSHDOWN is false.

All other test methods remain largely unchanged.
