+++ plugin/trino-postgresql/src/test/java/io/trino/plugin/postgresql/TestPostgreSqlConnectorTest.java
@@ -695,167 +695,6 @@ public class TestPostgreSqlConnectorTest
         }
     }
 
-    @Test
-    public void testOrPredicatePushdown()
-    {
-        assertThat(query("SELECT * FROM nation WHERE nationkey != 3 OR regionkey = 4")).isFullyPushedDown();
-        assertThat(query("SELECT * FROM nation WHERE nationkey != 3 OR regionkey != 4")).isFullyPushedDown();
-        assertThat(query("SELECT * FROM nation WHERE name = 'ALGERIA' OR regionkey = 4")).isFullyPushedDown();
-        assertThat(query("SELECT * FROM nation WHERE name IS NULL OR regionkey = 4")).isFullyPushedDown();
-        assertThat(query("SELECT * FROM nation WHERE name = NULL OR regionkey = 4")).isNotFullyPushedDown(FilterNode.class); // TODO `name = NULL` should be eliminated by the engine
-    }
-
-    @Test
-    public void testLikePredicatePushdown()
-    {
-        assertThat(query("SELECT nationkey FROM nation WHERE name LIKE '%A%'"))
-                .isFullyPushedDown();
-
-        try (TestTable table = new TestTable(
-                getQueryRunner()::execute,
-                "test_like_predicate_pushdown",
-                "(id integer, a_varchar varchar(1))",
-                List.of(
-                        "1, 'A'",
-                        "2, 'a'",
-                        "3, 'B'",
-                        "4, 'ą'",
-                        "5, 'Ą'"))) {
-            assertThat(query("SELECT id FROM " + table.getName() + " WHERE a_varchar LIKE '%A%'"))
-                    .isFullyPushedDown();
-            assertThat(query("SELECT id FROM " + table.getName() + " WHERE a_varchar LIKE '%ą%'"))
-                    .isFullyPushedDown();
-        }
-    }
-
-    @Test
-    public void testLikeWithEscapePredicatePushdown()
-    {
-        assertThat(query("SELECT nationkey FROM nation WHERE name LIKE '%A%' ESCAPE '\\'"))
-                .isFullyPushedDown();
-
-        try (TestTable table = new TestTable(
-                getQueryRunner()::execute,
-                "test_like_with_escape_predicate_pushdown",
-                "(id integer, a_varchar varchar(4))",
-                List.of(
-                        "1, 'A%b'",
-                        "2, 'Asth'",
-                        "3, 'ą%b'",
-                        "4, 'ąsth'"))) {
-            assertThat(query("SELECT id FROM " + table.getName() + " WHERE a_varchar LIKE '%A\\%%' ESCAPE '\\'"))
-                    .isFullyPushedDown();
-            assertThat(query("SELECT id FROM " + table.getName() + " WHERE a_varchar LIKE '%ą\\%%' ESCAPE '\\'"))
-                    .isFullyPushedDown();
-        }
-    }
-
-    @Test
-    public void testIsNullPredicatePushdown()
-    {
-        assertThat(query("SELECT nationkey FROM nation WHERE name IS NULL")).isFullyPushedDown();
-        assertThat(query("SELECT nationkey FROM nation WHERE name IS NULL OR regionkey = 4")).isFullyPushedDown();
-
-        try (TestTable table = new TestTable(
-                getQueryRunner()::execute,
-                "test_is_null_predicate_pushdown",
-                "(a_int integer, a_varchar varchar(1))",
-                List.of(
-                        "1, 'A'",
-                        "2, 'B'",
-                        "1, NULL",
-                        "2, NULL"))) {
-            assertThat(query("SELECT a_int FROM " + table.getName() + " WHERE a_varchar IS NULL OR a_int = 1")).isFullyPushedDown();
-        }
-    }
-
-    @Test
-    public void testIsNotNullPredicatePushdown()
-    {
-        assertThat(query("SELECT nationkey FROM nation WHERE name IS NOT NULL OR regionkey = 4")).isFullyPushedDown();
-
-        try (TestTable table = new TestTable(
-                getQueryRunner()::execute,
-                "test_is_not_null_predicate_pushdown",
-                "(a_int integer, a_varchar varchar(1))",
-                List.of(
-                        "1, 'A'",
-                        "2, 'B'",
-                        "1, NULL",
-                        "2, NULL"))) {
-            assertThat(query("SELECT a_int FROM " + table.getName() + " WHERE a_varchar IS NOT NULL OR a_int = 1")).isFullyPushedDown();
-        }
-    }
-
-    @Test
-    public void testNullIfPredicatePushdown()
-    {
-        assertThat(query("SELECT nationkey FROM nation WHERE NULLIF(name, 'ALGERIA') IS NULL"))
-                .matches("VALUES BIGINT '0'")
-                .isFullyPushedDown();
-
-        assertThat(query("SELECT name FROM nation WHERE NULLIF(nationkey, 0) IS NULL"))
-                .matches("VALUES CAST('ALGERIA' AS varchar(25))")
-                .isFullyPushedDown();
-
-        assertThat(query("SELECT nationkey FROM nation WHERE NULLIF(name, 'Algeria') IS NULL"))
-                .returnsEmptyResult()
-                .isFullyPushedDown();
-
-        // NULLIF returns the first argument because arguments aren't the same
-        assertThat(query("SELECT nationkey FROM nation WHERE NULLIF(name, 'Name not found') = name"))
-                .matches("SELECT nationkey FROM nation")
-                .isFullyPushedDown();
-    }
-
-    @Test
-    public void testNotExpressionPushdown()
-    {
-        assertThat(query("SELECT nationkey FROM nation WHERE NOT(name LIKE '%A%' ESCAPE '\\')")).isFullyPushedDown();
-
-        try (TestTable table = new TestTable(
-                getQueryRunner()::execute,
-                "test_is_not_predicate_pushdown",
-                "(a_int integer, a_varchar varchar(2))",
-                List.of(
-                        "1, 'Aa'",
-                        "2, 'Bb'",
-                        "1, NULL",
-                        "2, NULL"))) {
-            assertThat(query("SELECT a_int FROM " + table.getName() + " WHERE NOT(a_varchar LIKE 'A%') OR a_int = 2")).isFullyPushedDown();
-            assertThat(query("SELECT a_int FROM " + table.getName() + " WHERE NOT(a_varchar LIKE 'A%' OR a_int = 2)")).isFullyPushedDown();
-        }
-    }
-
-    @Test
-    public void testInPredicatePushdown()
-    {
-        try (TestTable table = new TestTable(
-                getQueryRunner()::execute,
-                "test_in_predicate_pushdown",
-                "(id varchar(1), id2 varchar(1))",
-                List.of(
-                        "'a', 'b'",
-                        "'b', 'c'",
-                        "'c', 'c'",
-                        "'d', 'd'",
-                        "'a', 'f'"))) {
-            // IN values cannot be represented as a domain
-            assertThat(query("SELECT id FROM " + table.getName() + " WHERE id IN ('a', id2)"))
-                    .isFullyPushedDown();
-
-            assertThat(query("SELECT id FROM " + table.getName() + " WHERE id IN ('a', 'b') OR id2 IN ('c', 'd')"))
-                    .isFullyPushedDown();
-
-            assertThat(query("SELECT id FROM " + table.getName() + " WHERE id IN ('a', 'B') OR id2 IN ('c', 'D')"))
-                    .isFullyPushedDown();
-
-            assertThat(query("SELECT id FROM " + table.getName() + " WHERE id IN ('a', 'B', NULL) OR id2 IN ('C', 'd')"))
-                    // NULL constant value is currently not pushed down
-                    .isNotFullyPushedDown(FilterNode.class);
-        }
-    }
-
     @Override
     protected void verifyDivisionByZeroFailure(Throwable e)
     {
